server/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-channels==0.7.0
channels-redis==4.0.0
idna==3.4
oauthlib==3.2.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
daphne==4.0.0
whitenoise